+++ src/components/FormSelectList/OptionboxView.vue
@@ -22,11 +22,12 @@
 <script>
 import { createUniqIdsMixin } from "vue-uniq-ids";
 import ValidationMixin from "../mixins/validation";
+import checkbox from "../mixins/checkbox";
 
 const uniqIdsMixin = createUniqIdsMixin();
 
 export default {
-  mixins: [uniqIdsMixin, ValidationMixin],
+  mixins: [uniqIdsMixin, ValidationMixin, checkbox],
   inheritAttrs: false,
   props: [
     "value",
@@ -48,21 +49,6 @@ export default {
       selected: []
     };
   },
-  computed: {
-    divClass() {
-      return this.toggle ? "custom-control custom-radio" : "form-check";
-    },
-    labelClass() {
-      return this.toggle ? "custom-control-label" : "form-check-label";
-    },
-    inputClass() {
-      return [
-        { [this.controlClass]: !!this.controlClass },
-        { "is-invalid": (this.validator && this.validator.errorCount) || this.error },
-        this.toggle ? "custom-control-input" : "form-check-input"
-      ];
-    }
-  },
   watch: {
     value(val) {
       this.selected = val;
@@ -73,32 +59,6 @@ export default {
   },
   mounted() {
     this.selected = this.value;
-  },
-  methods: {
-    getOptionValue(option) {
-      return option[this.optionValue || "value"];
-    },
-    getOptionContent(option) {
-      return option[this.optionContent || "content"];
-    },
-    getOptionAriaLabel(option) {
-      let ariaLabel = "";
-      if (this.optionsExtra?.length) {
-        const optionExtra = this.optionsExtra.find(
-          (extra) => extra.hasOwnProperty(this.optionValue) && extra[this.optionValue] === option[this.optionValue]
-        );
-        if (optionExtra) {
-          ariaLabel = optionExtra[this.optionAriaLabel || "ariaLabel"] ?? "";
-        }
-      } else {
-        ariaLabel = option[this.optionAriaLabel || "ariaLabel"] ?? "";
-      }
-      return !ariaLabel || ariaLabel === "" ? this.getOptionContent(option) : ariaLabel;
-    },
-    getOptionId(option, index) {
-      return `${this.name}-${this.getOptionValue(option)}-${index}`;
-    }
   }
 };
-
 </script>
